design: 공유하기 UI 변경
