+++ src/pages/Petition/PetitionContents/index.tsx
@@ -9,12 +9,15 @@ import {
   CurrentAgreements,
   PetitionDescription,
   ContentWrap,
+  SharingPetition,
 } from './styles'
 import AgreementList from './AgreementList'
 import AgreementForm from './AgreementForm'
 import { useDisclosure } from '@chakra-ui/react'
 import NeedLoginModal from '@components/NeedLoginModal'
 import { getDay } from '@utils/getTime'
+import { RiKakaoTalkFill, RiFacebookFill } from 'react-icons/ri'
+import { IoMdAlbums } from 'react-icons/io'
 
 interface IProps {
   petitionURL: string
@@ -27,6 +30,45 @@ const PetitionContents = ({ petitionURL, petitionId }: IProps): JSX.Element => {
     AnswerContent | undefined
   >()
   const { isOpen, onClose } = useDisclosure()
+
+  const clip = () => {
+    let url = ''
+    const textarea = document.createElement('textarea')
+    document.body.appendChild(textarea)
+    url = String(document.querySelector('.url')?.textContent)
+    textarea.value = url
+    textarea.select()
+    document.execCommand('copy')
+    document.body.removeChild(textarea)
+    alert('URL이 복사되었습니다.')
+  }
+
+  const share = (sns: string) => {
+    alert('구현중!')
+    // const thisUrl = `https://www.gist-petition.com/${petitionURL}`
+    // if (sns == 'facebook') {
+    //   const url =
+    //     'http://www.facebook.com/sharer/sharer.php?u=' +
+    //     encodeURIComponent(thisUrl)
+    //   window.open(url, '', 'width=486, height=286')
+    // } else if (sns == 'kakaotalk') {
+    //   window.Kakao.init(process.env.REACT_APP_KAKAO_JAVASCRIPT_KEY)
+    //   window.Kakao.Link.createDefaultButton({
+    //     container: '#btnKakao',
+    //     objectType: '',
+    //     content: {
+    //       title: 'GIST 청원',
+    //       description: '',
+    //       imageUrl: thisUrl,
+    //       link: {
+    //         mobileWebUrl: thisUrl,
+    //         webUrl: thisUrl,
+    //       },
+    //     },
+    //   })
+    // }
+  }
+
   useEffect(() => {
     const getPetitionInformation = async (petitionURL: string) => {
       const response = await getPetitionById(petitionURL)
@@ -97,6 +139,56 @@ const PetitionContents = ({ petitionURL, petitionId }: IProps): JSX.Element => {
             </Stack>
           ) : (
             <div>
+              <SharingPetition>
+                <div className="share-btns">
+                  <div>공유하기</div>
+                  <ul className="sns">
+                    <li className="kakaotalk">
+                      <a
+                        href="#n"
+                        id="btnKakao"
+                        onClick={() => {
+                          share('kakaotalk')
+                          return false
+                        }}
+                        className="kakaotalk"
+                        target="_self"
+                        title="카카오톡 새창열림"
+                      >
+                        <RiKakaoTalkFill />
+                      </a>
+                    </li>
+                    <li className="facebook">
+                      <a
+                        href="#n"
+                        onClick={() => {
+                          share('facebook')
+                          return false
+                        }}
+                        className="facebook"
+                        target="_self"
+                        title="페이스북 새창열림"
+                      >
+                        <RiFacebookFill />
+                      </a>
+                    </li>
+                  </ul>
+                </div>
+                <div className="share-url">
+                  <div className="url-box">
+                    <div>URL</div>
+                    <div className="url">
+                      https://www.gist-petition.com/
+                      {petitionURL}
+                    </div>
+                  </div>
+                  <div className="copy-btn">
+                    <button onClick={clip}>
+                      <IoMdAlbums />
+                    </button>
+                  </div>
+                </div>
+              </SharingPetition>
               <Stack>
                 <Text
                   textAlign={'left'}
